+++ packages/react/src/components/PlanDisplay/src/components/PlanDisplayBullet/index.ts
@@ -1,0 +1,5 @@
+import { PlanDisplayBullet } from "./PlanDisplayBullet";
+
+export { PlanDisplayBullet };
+
+export default PlanDisplayBullet;