+++ packages/react/src/components/PlanDisplay/src/components/PlanDisplayCard/planDisplayCard.spec.tsx
@@ -1,0 +1,27 @@
+import React from "react";
+import { render, screen } from "@testing-library/react";
+
+import { PlanDisplayCard } from "./PlanDisplayCard";
+import { PlanDisplayCardProps } from "./planDisplayCard.types";
+
+const bodyChildren = <div>Body content</div>;
+
+const makeSut = (rest: Omit<PlanDisplayCardProps, "children">) => {
+  render(<PlanDisplayCard {...rest}>{bodyChildren}</PlanDisplayCard>);
+};
+
+describe("GIVEN <PlanDisplayCard />", () => {
+  describe("WHEN rendered", () => {
+    it("SHOULD render children correctly", () => {
+      makeSut({});
+
+      expect(screen.getByText("Body content")).toBeDefined();
+    });
+
+    it("SHOULD render highlighted state correctly", () => {
+      makeSut({ highlighted: true });
+
+      expect(screen.getByText("Body content")).toBeDefined();
+    });
+  });
+});