+++ packages/react/src/components/PlanDisplay/src/components/PlanDisplayCard/planDisplayCard.types.ts
@@ -1,0 +1,6 @@
+import { PropsWithChildren } from "react";
+
+type CardProps = {
+  highlighted?: boolean;
+};
+export type PlanDisplayCardProps = PropsWithChildren<CardProps>;