+++ packages/react/src/components/PlanDisplay/src/components/PlanDisplayCardContent/PlanDisplayCardContent.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { Box, Card } from "@nimbus-ds/components";
+import { PlanDisplayCardContentProps } from "./planDisplayCardContent.types";
+import PlanDisplayCardSpacing from "../PlanDisplayCardSpacing";
+
+const PlanDisplayCardContent: React.FC<PlanDisplayCardContentProps> = ({
+  children,
+}) => (
+  <Card.Body>
+    <PlanDisplayCardSpacing />
+
+    <Box display="flex" flexDirection="column" gap="1" py="4">
+      {children}
+    </Box>
+  </Card.Body>
+);
+
+export { PlanDisplayCardContent };