+++ packages/react/src/components/PlanDisplay/src/components/PlanDisplayCardContent/index.ts
@@ -1,0 +1,5 @@
+import { PlanDisplayCardContent } from "./PlanDisplayCardContent";
+
+export { PlanDisplayCardContent };
+
+export default PlanDisplayCardContent;