+++ packages/react/src/components/PlanDisplay/src/components/PlanDisplayCardHeader/index.ts
@@ -1,0 +1,5 @@
+import { PlanDisplayCardHeader } from "./PlanDisplayCardHeader";
+
+export { PlanDisplayCardHeader };
+
+export default PlanDisplayCardHeader;