+++ packages/react/src/components/PlanDisplay/src/components/PlanDisplayCardHeader/planDisplayCardHeader.types.ts
@@ -1,0 +1,7 @@
+import { PropsWithChildren, ReactNode } from "react";
+
+type CardHeaderProps = {
+  subtitle: string;
+  title: ReactNode;
+};
+export type PlanDisplayCardHeaderProps = PropsWithChildren<CardHeaderProps>;