+++ packages/react/src/components/PlanDisplay/src/components/PlanDisplayCardSpacing/planDisplayCardSpacing.types.ts
@@ -1,0 +1,6 @@
+import { BoxProps } from "@nimbus-ds/components";
+
+export type PlanDisplayCardSpacingProps = Omit<
+  BoxProps,
+  "children" | "borderTopWidth" | "borderColor" | "borderStyle" | "my"
+>;