+++ packages/react/src/components/PlanDisplay/src/index.ts
@@ -1,0 +1,5 @@
+import { PlanDisplay } from "./PlanDisplay";
+
+export { PlanDisplay } from "./PlanDisplay";
+export type { PlanDisplayProps } from "./planDisplay.types";
+export default PlanDisplay;